backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
pillow
pytesseract
